--- Sigma/.settings/org.eclipse.core.resources.prefs
@@ -1,2 +1,0 @@
-eclipse.preferences.version=1
-encoding//src/main/webapp/WEB-INF/views/home.jsp=UTF-8
