+++ Sigma/src/main/java/kr/co/controller/boardController.java
@@ -14,9 +14,11 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import kr.co.domain.PageTO;
+import kr.co.domain.SPageTO;
 import kr.co.domain.boardVO;
 import kr.co.service.boardService;
 import kr.co.service.replyService;
+import kr.co.service.sboardService;
 
 @Controller
 @RequestMapping("/board")
@@ -26,42 +28,45 @@ public class boardController {
 	
 	@Autowired
 	private replyService rservice;
-
-	@RequestMapping(value = "boardFR/list", method = RequestMethod.GET)
+	
+	@Autowired
+	private sboardService sbService;
+	
+	@RequestMapping(value = "/main/list/boardlist", method = RequestMethod.GET)
 	public void boardFRList(PageTO<boardVO> to, Model model) {
 		
 		PageTO<boardVO> dbTO = bservice.pageList(to);
 		model.addAttribute("dbTO", dbTO);
 	}
 	
-	@RequestMapping(value = "/boardFR/insert", method = RequestMethod.GET)
+	@RequestMapping(value = "/main/list/boardinsert", method = RequestMethod.GET)
 	public void boardFRinsertUI() {
 	}
 	
-	@RequestMapping(value = "/boardFR/insert", method = RequestMethod.POST)
+	@RequestMapping(value = "/main/list/boardinsert", method = RequestMethod.POST)
 	public String boardFRinsert(boardVO vo) {
 		bservice.boardFRinsert(vo);
-		return "redirect:/mainboard";
+		return "redirect:/board/main/list/boardlist";
 	}
 	
-	@RequestMapping(value = "/boardFR/read")
+	@RequestMapping(value = "/main/list/boardread")
 	public void boardFRread(boardVO vo, Model model , PageTO<boardVO> to) {
 		boardVO readvo  = bservice.boardFRread(vo);
 		model.addAttribute("readvo", readvo);
 		model.addAttribute("to", to);
 	}
 	
-	@RequestMapping(value = "/boardFR/update")
+	@RequestMapping(value = "/main/list/boardupdate")
 	public void boardFRupdateUI(int num,Model model,PageTO<boardVO> to) {
 		boardVO updatevo = bservice.boardFRupdateUI(num);
 		model.addAttribute("updatevo", updatevo);
 		model.addAttribute("to", to);
 	}
 	
-	@RequestMapping(value = "/boardFR/update", method = RequestMethod.POST)
+	@RequestMapping(value = "/main/list/boardupdate", method = RequestMethod.POST)
 	public String boardFRupdate(boardVO vo , PageTO<boardVO> to) {
 		bservice.boardFRupdate(vo);
-		return "redirect:/board/boardFR/read?num=" + vo.getNum()+"&curPage="+to.getCurPage()+"&perPage="+to.getPerPage();
+		return "redirect:/board/main/list/boardread?num=" + vo.getNum()+"&curPage="+to.getCurPage()+"&perPage="+to.getPerPage();
 		
 	}
 	
@@ -69,7 +74,7 @@ public class boardController {
 	public String boardFRdelete(int num , PageTO<boardVO> to) throws Exception {
 		rservice.deleteAll(num);
 		bservice.boardFRdelete(num);
-		return "redirect:/mainboard?curPage="+to.getCurPage()+"&perPage="+to.getPerPage();
+		return "redirect:/board/main/list/boardlist?curPage="+to.getCurPage()+"&perPage="+to.getPerPage();
 	}
 	
 	@ResponseBody
@@ -80,11 +85,53 @@ public class boardController {
 		return (amount-1)/perPage+1;
 	}
 	
-	@RequestMapping(value = "/boardFR/mainboard", method = RequestMethod.GET)
-	public void mainboard(PageTO<boardVO> to, Model model) {
+	//==========================================slist======================================================
+	@RequestMapping("/main/slist/searchboardlist")
+	public void list(SPageTO sto, Model model) {
+		SPageTO dbSTO = sbService.list(sto);
+		model.addAttribute("to", dbSTO); 
+	}
+	
+	@RequestMapping(value = "/main/slist/searchboardread")
+	public void read(Model model, int num, SPageTO sto) {
+		boardVO svo = sbService.read(num);
 		
-		PageTO<boardVO> dbTO = bservice.pageList(to);
-		model.addAttribute("dbTO", dbTO);
+		model.addAttribute("vo", svo);
+		model.addAttribute("to", sto);
+	}
+	
+	@ResponseBody
+	@RequestMapping(value = "/searchMainboardDel", method = RequestMethod.POST)
+	public void delete(int num) {
+		rservice.deleteAll(num);
+	    sbService.del(num);
+	}
+	
+	@RequestMapping(value="/main/slist/searchboardupdate", method = RequestMethod.GET) 
+	public void	updateUI(Model model, SPageTO sto, int num) {
+		boardVO vo = sbService.updateUI(num); 
+		model.addAttribute("vo", vo);
+		model.addAttribute("to", sto); 
+	}
+
+	@RequestMapping(value="/main/slist/searchboardupdate", method = RequestMethod.POST)
+	public String update(boardVO vo, SPageTO sto) { 
 		
+		sbService.update(vo);
+		
+		StringBuffer sb = new StringBuffer();
+		sb.append("redirect:/board/main/slist/searchboardread?num=");
+		sb.append(vo.getNum());
+		sb.append("&CurPage=");
+		sb.append(sto.getCurPage());
+		sb.append("&perPage=");
+		sb.append(sto.getPerPage());
+		sb.append("&searchType=");
+		sb.append(sto.getSearchType());
+		sb.append("&keyword=");
+		sb.append(sto.getKeyword());
+		
+		return sb.toString();
+//		return "redirect:/sboard/read?bno="+vo.getBno()+"&curPage="+sto.getCurPage()+"&perPage"+sto.getPerPage()+"&searchType="+sto.getSearchType()+"&keyword="+sto.getKeyword();
 	}
 }
