--- Sigma/src/main/webapp/WEB-INF/views/NewGameInsert.jsp
@@ -1,47 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-	pageEncoding="UTF-8"%>
-<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
-<%@ taglib uri="http://java.sun.com/jsp/jstl/fmt" prefix="fmt"%>
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>메인 홈페이지</title>
-<meta name="viewport" content="width=device-width,initial-scale=1">
-
-<!-- Latest compiled and minified CSS -->
-<link rel="stylesheet"
-	href="https://stackpath.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-<!-- Font Awesome -->
-<link rel="stylesheet"
-	href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
-
-
-<link href="https://fonts.googleapis.com/icon?family=Material+Icons"
-	rel="stylesheet">
-
-<script
-	src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
-<!-- Latest compiled and minified JavaScript -->
-<script
-	src="https://stackpath.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
-
-<style type="text/css">
-html, body {
-	background: linear-gradient(60deg, orange, black ,#5D5D5D,gray,black,orange);
-	margin: auto;
-	width: auto;
-	height: auto;
-}
-
-</style>
-</head>
-<body>
-	<!-- 헤더 -->
-	<jsp:include page="/WEB-INF/views/main/header.jsp"></jsp:include>
-	<!-- 바디 -->
-	<jsp:include page="/WEB-INF/views/board/boardNG/insert.jsp"></jsp:include>
-	<!-- 풋터 -->
-	<jsp:include page="/WEB-INF/views/main/footer.jsp"></jsp:include>
-</body>
-</html>
