+++ Sigma/src/main/webapp/WEB-INF/views/board/boardFR/list.jsp
@@ -66,7 +66,7 @@
 			<c:forEach items="${dbTO.list}" var="vo">
 				<tr>
 					<td>${vo.num}</td>
-					<td><a href="/mainboardread?num=${vo.num}&curPage=${dbTO.curPage}&perPage=${dbTO.perPage}">${vo.title}</a></td>
+					<td><a href="/board/main/list/boardread?num=${vo.num}&curPage=${dbTO.curPage}&perPage=${dbTO.perPage}">${vo.title}</a></td>
 					<td>${vo.writer}</td>
 					<td>${vo.updatedate}</td>
 					<td>${vo.viewcnt}</td>
@@ -75,7 +75,7 @@
 		</tbody>
 	</table>
 	
-	<a href="/NewGameInsert" class="btn btn-primary" style="float: right;">글쓰기</a>
+	<a href="/board/main/list/boardinsert" class="btn btn-primary" style="float: right;">글쓰기</a>
 <%-- 		<div class="row text-center" style="text-align: center">
 				<ul class="pagination"> <!-- ul에 pagination 클래스를 주면 예쁘다 -->
 				
@@ -117,19 +117,19 @@
 
 						<c:if test="${dbTO.curPage>1}">
 							<li><a
-								href="/mainboard?curPage=${dbTO.curPage-1}&perPage=${dbTO.perPage}">&laquo;</a></li>
+								href="/board/main/list/boardlist?curPage=${dbTO.curPage-1}&perPage=${dbTO.perPage}">&laquo;</a></li>
 						</c:if>
 						<!-- 주소창에서 perPage값을 조절하면서 확인할것 -->
 
 						<c:forEach begin="${dbTO.bpn}" end="${dbTO.spn}" var="idx">
 							<li class="${dbTO.curPage == idx?'active':''}"><a
-								href="/mainboard?curPage=${idx}&perPage=${dbTO.perPage}">${idx}</a></li>
+								href="/board/main/list/boardlist?curPage=${idx}&perPage=${dbTO.perPage}">${idx}</a></li>
 							<!-- li에 클래스를 active로 주면 현재 페이지에 색이 들어간다 -->
 						</c:forEach>
 
 						<c:if test="${dbTO.curPage<dbTO.totalPage}">
 							<li><a
-								href="/mainboard?curPage=${dbTO.curPage+1}&perPage=${dbTO.perPage}">&raquo;</a></li>
+								href="/board/main/list/boardlist?curPage=${dbTO.curPage+1}&perPage=${dbTO.perPage}">&raquo;</a></li>
 						</c:if>
 
 					</ul>
@@ -171,9 +171,9 @@
 				            dataType:'text',
 				            success: function(totalPage) {
 				               if(${dbTO.curPage}>totalPage){
-				                  location.assign("/mainboard?perPage="+perPage+"&curPage="+totalPage);
+				                  location.assign("/board/main/list/boardlist?perPage="+perPage+"&curPage="+totalPage);
 				               }else {
-				                  location.assign("/mainboard?perPage="+perPage+"&curPage=${dbTO.curPage}");
+				                  location.assign("/board/main/list/boardlist?perPage="+perPage+"&curPage=${dbTO.curPage}");
 				               }
 				            }
 				         });
@@ -182,7 +182,7 @@
 					   $("#searchBtn").on("click", function() {
 							var searchType = $("#searchSel option:selected").val();
 							var keyword = $("#keyword").val();
-							var url = "/searchMainboard?searchType="+searchType+"&keyword="+keyword;
+							var url = "/board/main/slist/searchboardlist?searchType="+searchType+"&keyword="+keyword;
 							window.open(url);
 						});
 				   });
