+++ Sigma/src/main/webapp/WEB-INF/views/board/boardFR/read.jsp
@@ -67,7 +67,7 @@
 				</div>
 				<div>
 				    <input type="button" value="수정" class="btn update btn-primary">
-					<input type="button" value="목록" class="btn btn-info" onclick="location.href='/mainboard?curPage=${to.curPage}&perPage=${to.perPage}'">
+					<input type="button" value="목록" class="btn btn-info" onclick="location.href='/board/main/list/boardlist?curPage=${to.curPage}&perPage=${to.perPage}'">
 					<input type="button" value="삭제" class="btn btn-warning" onclick="location.href='/board/boardFR/delete?num=${readvo.num}&curPage=${to.curPage}&perPage=${to.perPage}'">
 				</div>
 			</div>
@@ -212,7 +212,7 @@
 		
 		/* boardUpdate 부분 */
 		$(".update").on("click",function(){
-			location.href="/mainboardupdate?num=${readvo.num}&curPage=${to.curPage}&perPage=${to.perPage}";
+			location.href="/board/main/list/boardupdate?num=${readvo.num}&curPage=${to.curPage}&perPage=${to.perPage}";
 		})
 		
 		/* replyReset 부분 */
