+++ Sigma/src/main/webapp/WEB-INF/views/board/boardNG/list.jsp
@@ -45,7 +45,7 @@
 	<br>
 <div class="bodymain">
 		<div class="list">
-<div class="container">
+			<div class="container">
 <%--   	<div class="row">
          <div class="col-xs-push-10 col-xs-2">
             <select id="perPageSel" class="form-control">
@@ -73,7 +73,7 @@
 			<c:forEach items="${dbTO.list}" var="vo">
 				<tr>
 					<td>${vo.num}</td>
-					<td><a href="/NewGameRead?num=${vo.num}&curPage=${dbTO.curPage}&perPage=${dbTO.perPage}">${vo.title}</a></td>
+					<td><a href="/board/main/newgame/NewGameRead?num=${vo.num}&curPage=${dbTO.curPage}&perPage=${dbTO.perPage}">${vo.title}</a></td>
 					<td>${vo.writer}</td>
 					<td>${vo.updatedate}</td>
 					<td>${vo.viewcnt}</td>
@@ -82,26 +82,26 @@
 		</tbody>
 	</table>
 	
-		<a href="/NewGameInsert" class="btn btn-primary" style="float: right;">글쓰기</a>
+		<a href="/board/main/newgame/NewGameInsert" class="btn btn-primary" style="float: right;">글쓰기</a>
 		
 		<div class="row text-center" style="text-align: center">
 				<ul class="pagination"> <!-- ul에 pagination 클래스를 주면 예쁘다 -->
 				
 					<c:if test="${dbTO.curPage>1}">
 						<li><a
-							href="/NewGame?curPage=${dbTO.curPage-1}&perPage=${dbTO.perPage}">&laquo;</a></li>
+							href="/board/main/newgame/NewGame?curPage=${dbTO.curPage-1}&perPage=${dbTO.perPage}">&laquo;</a></li>
 					</c:if>
 						<!-- 주소창에서 perPage값을 조절하면서 확인할것 -->
 
 					<c:forEach begin="${dbTO.bpn}" end="${dbTO.spn}" var="idx">
 						<li class="${dbTO.curPage == idx?'active':''}"><a
-							href="/NewGame?curPage=${idx}&perPage=${dbTO.perPage}">${idx}</a></li>
+							href="/board/main/newgame/NewGame?curPage=${idx}&perPage=${dbTO.perPage}">${idx}</a></li>
 						<!-- li에 클래스를 active로 주면 현재 페이지에 색이 들어간다 -->
 					</c:forEach>
 
 					<c:if test="${dbTO.curPage<dbTO.totalPage}">
 						<li><a
-							href="/NewGame?curPage=${dbTO.curPage+1}&perPage=${dbTO.perPage}">&raquo;</a></li>
+							href="/board/main/newgame/NewGame?curPage=${dbTO.curPage+1}&perPage=${dbTO.perPage}">&raquo;</a></li>
 					</c:if>
 
 				</ul>
@@ -130,7 +130,9 @@
 			</div>
 			</div>
 			<br><br>
+			</div>
 			
+			</div>
 			<script type="text/javascript">
 			   $(document).ready(function() {
 /* 				      $("#perPageSel").change(function() {
