+++ Sigma/src/main/webapp/WEB-INF/views/board/main/list/boardinsert.jsp
