+++ Sigma/src/main/webapp/WEB-INF/views/board/main/newgame/NewGameInsert.jsp
@@ -25,7 +25,7 @@
 <!-- Latest compiled and minified JavaScript -->
 <script
 	src="https://stackpath.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
-
+<link rel="stylesheet" type="text/css" href="/resources/css/main.css">
 <style type="text/css">
 html, body {
 	background: linear-gradient(60deg, orange, black ,#5D5D5D,gray,black,orange);
@@ -34,15 +34,13 @@ html, body {
 	height: auto;
 }
 
-
-
 </style>
 </head>
 <body>
 	<!-- 헤더 -->
 	<jsp:include page="/WEB-INF/views/main/header.jsp"></jsp:include>
 	<!-- 바디 -->
-	<jsp:include page="/WEB-INF/views/board/boardEV/list.jsp"></jsp:include>
+	<jsp:include page="/WEB-INF/views/board/boardNG/insert.jsp"></jsp:include>
 	<!-- 풋터 -->
 	<jsp:include page="/WEB-INF/views/main/footer.jsp"></jsp:include>
 </body>
