+++ Sigma/src/main/webapp/WEB-INF/views/board/main/slist/searchboardupdate.jsp
