+++ Sigma/src/main/webapp/WEB-INF/views/compManage/refund/refundList.jsp
@@ -39,8 +39,8 @@
 				<c:if test="${!empty list}">
 					<c:forEach items="${list}" var="refundlist">
 					<tr>
-						<td>${refundlist.id}</td>
-						<%-- <td><a href="/compManage/refund/refundRead?id=${login.id}&num=${refundlist.num}">${refundlist.title}</a></td> --%>
+						<td>${refundlist.id}</td>		
+						<td><a href="/compManage/refund/refundRead?id=${login.id}&num=${refundlist.num}">${refundlist.title}</a></td>
 						<td>${refundlist.updatedate}</td>
 						<td>${refundlist.status}</td>
 					</tr>
