+++ Sigma/src/main/webapp/WEB-INF/views/gameDetail/main/maincategory.jsp
