+++ Sigma/src/main/webapp/WEB-INF/views/gameDetail/main/maincategoryupdate.jsp
