--- Sigma/src/main/webapp/WEB-INF/views/home.jsp
@@ -1,22 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
-<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
-<%@ taglib uri="http://java.sun.com/jsp/jstl/fmt" prefix="fmt"%>
-<html>
-<head>
-	<title>Home</title>
-</head>
-<body>
-<h3>
-	 Sigma main home page.  
-</h3>
-
-<P>Please click the follow link.</P>
-<br/>
-
-<a href="#">게임추천</a>
-<a href="/gameDetail/inform/list?category=all">카테고리별</a>
-<a href="/board/boardFR/list">자유게시판</a>
-<a href="/compManage/compInform/read?id=idid">판매자 페이지</a>
-</body>
-</html>
