+++ Sigma/src/main/webapp/WEB-INF/views/myPage/myPage/mypage.jsp
@@ -11,7 +11,11 @@
 <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
 <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
-
+<style type="text/css">
+	.row{
+	color: white;
+	}
+</style>
 </head>
 <body>
 	<div class="container">
